divided flipper.rb on 2 files, replaced send with public_send in features service
